Date: 14.10.25
subject: semi-working count stepping algorithm

Desc:
I edited the StepCounterZC.kt file so the algorithm counts
steps when the phone's screen is faced towards the user or placed in the user's pocket.

this version will serve as a rollback version in case of a safe starting place to return to.
what is needed to be done from this point forward:
1. reduce significantly the count of hand tilts and shakes as valid steps.
2. allow step counting if the phone is horizontal when walking.

committer: Maor Mordo